Move to uk.ac.ebi.quickgo.ontology.model.graph
