+++ ontology-rest/src/main/java/uk/ac/ebi/quickgo/ontology/model/graph/AncestorEdge.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.quickgo.ontology.model;
+package uk.ac.ebi.quickgo.ontology.model.graph;
 
 /**
  * Data structure for the Ancestor Graph edges.
